pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.80.0",
    "idna>=3.11",
    "numpy>=1.24.0",
    "openai-whisper>=20231117",
    "piper-tts>=1.2.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "urllib3>=2.6.3",
]
